API Reclass : - report.form.dataList.master.getInstitutionType -> report.form.dataList.master.institutionType.getAll - report.form.dataList.master.getPerson -> report.form.dataList.master.person.getAll - report.form.dataList.master.getWorkStructure -> report.form.dataList.master.workStructure.getAll
