+++ Programming/.LaravelCore/app/Http/Controllers/Application/BackEnd/System/Report/Engines/form/dataList/master/institutionType/getAll/v1/JSONRequestSchema.json
@@ -1,6 +1,6 @@
 {
 "title": "JSON Request Schema",
-"description": "JSON Request Schema of report.form.dataList.master.getInstitutionType (version 1)",
+"description": "JSON Request Schema of report.form.dataList.master.institutionType.getAll (version 1)",
 "type": "object",
 "properties": {
     "metadata": {
@@ -11,7 +11,7 @@
                 "properties": {
                     "key": {
                         "type": "string",
-                        "enum": [ "report.form.dataList.master.getInstitutionType" ]
+                        "enum": [ "report.form.dataList.master.institutionType.getAll" ]
                         },
                     "version": {
                         "type": [ "integer", "string" ],
