+++ Programming/.LaravelCore/app/Http/Controllers/Application/BackEnd/System/Report/Engines/form/dataList/master/workStructure/getAll/v1/JSONRequestSchema.json
@@ -1,6 +1,6 @@
 {
     "title": "JSON Request Schema",
-    "description": "JSON Request Schema of report.form.dataList.master.getWorkStructure (version 1)",
+    "description": "JSON Request Schema of report.form.dataList.master.workStructure.getAll (version 1)",
     "type": "object",
     "properties": {
         "metadata": {
@@ -12,7 +12,7 @@
                         "key": {
                             "type": "string",
                             "enum": [
-                                "report.form.dataList.master.getWorkStructure"
+                                "report.form.dataList.master.workStructure.getAll"
                             ]
                         },
                         "version": {
