+++ Programming/.LaravelCore/app/Http/Controllers/Application/BackEnd/System/Report/Engines/form/dataList/master/workStructure/getAll/v1/getAll.php
@@ -3,29 +3,29 @@
 /*
 +----------------------------------------------------------------------------------------------------------------------------------+
 | ▪ Category   : API Engine Controller                                                                                             |
-| ▪ Name Space : \App\Http\Controllers\Application\BackEnd\System\Report\Engines\form\dataList\master\getInstitutionType           |
+| ▪ Name Space : \App\Http\Controllers\Application\BackEnd\System\Report\Engines\form\dataList\master\workStructure\getAll         |
 |                                                                                                                                  |
-| ▪ Copyleft 🄯 2026 Zheta (teguhpjs@gmail.com)                                                                                     |
+| ▪ Copyleft 🄯 2026 Wisnu (wisnu.wirayuda01@gmail.com)                                                                             |
 +----------------------------------------------------------------------------------------------------------------------------------+
 */
 
-namespace App\Http\Controllers\Application\BackEnd\System\Report\Engines\form\dataList\master\getInstitutionType\v1
+namespace App\Http\Controllers\Application\BackEnd\System\Report\Engines\form\dataList\master\workStructure\getAll\v1
     {
     /*
     +------------------------------------------------------------------------------------------------------------------------------+
-    | ▪ Class Name  : getInstitutionType                                                                                           |
-    | ▪ Description : Menangani API report.form.dataList.master.getInstitutionType Version 1                                       |
+    | ▪ Class Name  : getAll                                                                                                       |
+    | ▪ Description : Menangani API report.form.dataList.master.workStructure.getAll Version 1                                     |
     +------------------------------------------------------------------------------------------------------------------------------+
     */
-    class getInstitutionType extends \App\Http\Controllers\Controller
+    class getAll extends \App\Http\Controllers\Controller
         {
         /*
         +--------------------------------------------------------------------------------------------------------------------------+
         | ▪ Method Name     : __construct                                                                                          |
         +--------------------------------------------------------------------------------------------------------------------------+
         | ▪ Version         : 1.0000.0000000                                                                                       |
-        | ▪ Last Update     : 2026-06-08                                                                                           |
-        | ▪ Create date     : 2026-06-08                                                                                           |
+        | ▪ Last Update     : 2026-08-06                                                                                           |
+        | ▪ Create date     : 2026-08-06                                                                                           |
         | ▪ Description     : System's Default Constructor                                                                         |
         +--------------------------------------------------------------------------------------------------------------------------+
         | ▪ Input Variable  :                                                                                                      |
@@ -44,8 +44,8 @@ namespace App\Http\Controllers\Application\BackEnd\System\Report\Engines\form\da
         | ▪ Method Name     : main                                                                                                 |
         +--------------------------------------------------------------------------------------------------------------------------+
         | ▪ Version         : 1.0000.0000000                                                                                       |
-        | ▪ Last Update     : 2026-06-08                                                                                           |
-        | ▪ Create date     : 2026-06-08                                                                                           |
+        | ▪ Last Update     : 2026-08-06                                                                                           |
+        | ▪ Create date     : 2026-08-06                                                                                           |
         | ▪ Description     : Fungsi Utama Engine                                                                                  |
         +--------------------------------------------------------------------------------------------------------------------------+
         | ▪ Input Variable  :                                                                                                      |
@@ -62,7 +62,7 @@ namespace App\Http\Controllers\Application\BackEnd\System\Report\Engines\form\da
             try {
                 $varSysDataProcess =
                     \App\Helpers\ZhtHelper\Logger\Helper_SystemLog::setLogOutputMethodProcessHeader($varUserSession, __CLASS__, __FUNCTION__,
-                        'Get Report Form - DataList - Institution Type (version 1)'
+                        'Get Report Form - DataList - Work Structure (version 1)'
                         );
 
                 try {
@@ -71,7 +71,7 @@ namespace App\Http\Controllers\Application\BackEnd\System\Report\Engines\form\da
                         if (!($varDataSend = 
                             \App\Helpers\ZhtHelper\System\BackEnd\Helper_API::getEngineDataSend_DataRead(
                                 $varUserSession,
-                                (new \App\Models\Database\SchData_OLTP_Master\General())->getReport_Form_DataList_InstitutionType(
+                                (new \App\Models\Database\SchData_OLTP_Master\General())->getReport_Form_DataList_Work_Structure(
                                     //-----[ System Parameter ]-----------------------------------------------------( START POINT )-----
                                         //---► userSession
                                             $varUserSession,
@@ -116,38 +116,28 @@ namespace App\Http\Controllers\Application\BackEnd\System\Report\Engines\form\da
                                                 'name',
                                                 $varData['parameter']['dataFilter']
                                                 ) ? (
-                                                    (!is_null($varData['parameter']['dataFilter']['name'])) 
-                                                        ? $varData['parameter']['dataFilter']['name'] 
-                                                        : null
-                                                        ) 
-                                                    : null
-                                            ),
-                                        //---► country_RefID
-                                            (\App\Helpers\ZhtHelper\General\Helper_Array::isKeyExist(
-                                                $varUserSession,
-                                                'country_RefID',
-                                                $varData['parameter']['dataFilter']
-                                                ) ? (
-                                                    (!is_null($varData['parameter']['dataFilter']['country_RefID'])) 
-                                                        ? $varData['parameter']['dataFilter']['country_RefID'] 
-                                                        : null
-                                                        ) 
-                                                    : null
-                                            ),
-                                        //---► countryName
-                                            (
-                                            \App\Helpers\ZhtHelper\General\Helper_Array::isKeyExist(
-                                                $varUserSession,
-                                                'countryName',
-                                                $varData['parameter']['dataFilter']
-                                                ) ? (
-                                                    (!is_null($varData['parameter']['dataFilter']['countryName']))
-                                                        ? $varData['parameter']['dataFilter']['countryName']
+                                                    (!is_null($varData['parameter']['dataFilter']['name']))
+                                                        ? $varData['parameter']['dataFilter']['name']
                                                         : null
                                                         )
                                                     : null
+                                            ),
+                                        //---► code
+                                            (\App\Helpers\ZhtHelper\General\Helper_Array::isKeyExist(
+                                                $varUserSession,
+                                                'code',
+                                                $varData['parameter']['dataFilter']
+                                                ) ? (
+                                                    (!is_null($varData['parameter']['dataFilter']['code']))
+                                                    ? $varData['parameter']['dataFilter']['code']
+                                                    : null
+                                                    )
+                                                    : null
                                             )
                                     //-----[ Filter Parameter ]-----------------------------------------------------(  END POINT  )-----
+
+                                    //-----[ Order Parameter ]------------------------------------------------------( START POINT )-----
+                                    //-----[ Order Parameter ]------------------------------------------------------(  END POINT  )-----
                                     ),
                                 FALSE
                                 )
