+++ Programming/.LaravelCore/app/Http/Controllers/Application/FrontEnd/SandBox/Examples_APICall/report/form/dataList/master/institutionType/getAll/v1/example.php
@@ -3,21 +3,22 @@
 /*
 +----------------------------------------------------------------------------------------------------------------------------------+
 | ▪ Category    : Example - API Call Controller                                                                                    |
-| ▪ Name Space  : \App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\report\form\dataList\master\getPerson\v1     |
-| ▪ API Key     : report.form.dataList.master.getPerson                                                                            |
+| ▪ Name Space  : \App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\report\form\dataList\master\institutionType  |
+|                 \getAll\v1                                                                                                       |
+| ▪ API Key     : report.form.dataList.master.institutionType.getAll                                                               |
 | ▪ API Version : 1                                                                                                                |
 |                                                                                                                                  |
 | ▪ Copyleft 🄯 2026 Zheta (teguhpjs@gmail.com)                                                                                     |
 +----------------------------------------------------------------------------------------------------------------------------------+
 */
-namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\report\form\dataList\master\getPerson\v1
+namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\report\form\dataList\master\institutionType\getAll\v1
     {
     class example extends \App\Http\Controllers\Controller
         {
         /*
         +--------------------------------------------------------------------------------------------------------------------------+
-        | ▪ Call URL        : http(s)://<HOST>/report.form.dataList.master.getPerson.v1_throughAPIGateway                          |
-        |                     ► http://172.28.0.4/report.form.dataList.master.getPerson.v1_throughAPIGateway                       |
+        | ▪ Call URL        : http(s)://<HOST>/report.form.dataList.master.institutionType.getAll.v1_throughAPIGateway             |
+        |                     ► http://172.28.0.4/report.form.dataList.master.institutionType.getAll.v1_throughAPIGateway          |
         +--------------------------------------------------------------------------------------------------------------------------+
         | ▪ Version         : 1.0000.0000000                                                                                       |
         | ▪ Last Update     : 2026-06-08                                                                                           |
@@ -38,7 +39,7 @@ namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\rep
                     //-----[ METADATA ]-----( START )-----
                         \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System(),
                         $varAPIWebToken, 
-                        'report.form.dataList.master.getPerson', 
+                        'report.form.dataList.master.institutionType.getAll', 
                         'latest',
                     //-----[ METADATA ]-----(  END  )-----
 
@@ -50,8 +51,9 @@ namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\rep
                                 'pageShow' => 1
                                 ],
                             'dataFilter' => [
-                                'name' => 'Teguh',                                //'Teguh',
-                                'citizenIdentityNumber' => '3174092710810006'     //'3174092710810006'
+                                'name' => NULL,             //'Perusahaan',
+                                'country_RefID' => NULL,    //20000000000078
+                                'countryName' => NULL       //'Indonesia'
                                 ],
                             ]
                         ]
@@ -66,8 +68,8 @@ namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\rep
 
         /*
         +--------------------------------------------------------------------------------------------------------------------------+
-        | ▪ Call URL        : http(s)://<HOST>/transaction.read.dataList.master.getPerson.v1_throughAPIGatewayJQuery               |
-        |                     ► http://172.28.0.4/transaction.read.dataList.master.getPerson.v1_throughAPIGatewayJQuery            |
+        | ▪ Call URL        : http(s)://<HOST>/transaction.read.dataList.master.institutionType.getAll.v1_throughAPIGatewayJQuery      |
+        |                     ► http://172.28.0.4/transaction.read.dataList.master.institutionType.getAll.v1_throughAPIGatewayJQuery   |
         +--------------------------------------------------------------------------------------------------------------------------+
         | ▪ Version         : 1.0000.0000000                                                                                       |
         | ▪ Last Update     : 2022-09-08                                                                                           |
@@ -97,7 +99,7 @@ namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\rep
                     //-----[ METADATA ]-----( START )-----
                         \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System(), 
                         $varAPIWebToken, 
-                        'transaction.read.dataList.master.getPerson', 
+                        'transaction.read.dataList.master.getAll', 
                         'latest',
                     //-----[ METADATA ]-----(  END  )-----
