+++ Programming/.LaravelCore/app/Http/Controllers/Application/FrontEnd/SandBox/Examples_APICall/report/form/dataList/master/workStructure/getAll/v1/example.php
@@ -3,20 +3,22 @@
 /*
 +----------------------------------------------------------------------------------------------------------------------------------+
 | ▪ Category    : Example - API Call Controller                                                                                    |
-| ▪ Name Space  : \App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\report\form\dataList\master\getWorkStructure\v1       |
-| ▪ API Key     : report.form.dataList.master.getWorkStructure                                                                              |
+| ▪ Name Space  : \App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\report\form\dataList\master\workStructure    |
+|                 \getAll\v1                                                                                                       |
+| ▪ API Key     : report.form.dataList.master.workStructure.getAll                                                                 |
 | ▪ API Version : 1                                                                                                                |
 |                                                                                                                                  |
-| ▪ Copyleft 🄯 2026 Wisnu (wisnu.wirayuda01@gmail.com)                                                                                     |
+| ▪ Copyleft 🄯 2026 Wisnu (wisnu.wirayuda01@gmail.com)                                                                             |
 +----------------------------------------------------------------------------------------------------------------------------------+
 */
-namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\report\form\dataList\master\getWorkStructure\v1 {
-    class example extends \App\Http\Controllers\Controller
+namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\report\form\dataList\master\workStructure\getAll\v1
     {
+    class example extends \App\Http\Controllers\Controller
+        {
         /*
         +--------------------------------------------------------------------------------------------------------------------------+
-        | ▪ Call URL        : http(s)://<HOST>/report.form.dataList.master.getWorkStructure.v1_throughAPIGateway                            |
-        |                     ► http://172.28.0.4/report.form.dataList.master.getWorkStructure.v1_throughAPIGateway                         |
+        | ▪ Call URL        : http(s)://<HOST>/report.form.dataList.master.workStructure.getAll.v1_throughAPIGateway               |
+        |                     ► http://172.28.0.4/report.form.dataList.master.workStructure.getAll.v1_throughAPIGateway            |
         +--------------------------------------------------------------------------------------------------------------------------+
         | ▪ Version         : 1.0000.0000000                                                                                       |
         | ▪ Last Update     : 2026-08-06                                                                                           |
@@ -24,49 +26,49 @@ namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\rep
         +--------------------------------------------------------------------------------------------------------------------------+
         */
         public function throughAPIGateway($varAPIWebToken)
-        {
+            {
             //-----[ PARAMETER SET ]-----
             if (!$varAPIWebToken) {
                 $varAPIWebToken =
                     \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System();
-            }
+                }
 
             //-----[ CORE ]-----
             $varData =
                 \App\Helpers\ZhtHelper\System\FrontEnd\Helper_APICall::setCallAPIGateway(
                     //-----[ METADATA ]-----( START )-----
-                    \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System(),
-                    $varAPIWebToken,
-                    'report.form.dataList.master.getWorkStructure',
-                    'latest',
+                        \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System(),
+                        $varAPIWebToken,
+                        'report.form.dataList.master.workStructure.getAll',
+                        'latest',
                     //-----[ METADATA ]-----(  END  )-----
 
                     //-----[ DATA ]---------( START )-----
-                    [
+                        [
                         'parameter' => [
                             'pagination' => [
                                 'pageSize' => 10,
                                 'pageShow' => 1
-                            ],
+                                ],
                             'dataFilter' => [
                                 'name' => NULL,     //'Bank'
                                 'code' => null   //'BCA'
-                            ],
+                                ],
+                            ]
                         ]
-                    ]
                     //-----[ DATA ]---------(  END  )-----
-                );
+                    );
 
             //-----[ DATA RETURN ]-----
             return
                 $varData;
-        }
+            }
 
 
         /*
         +--------------------------------------------------------------------------------------------------------------------------+
-        | ▪ Call URL        : http(s)://<HOST>/transaction.read.dataList.master.getWorkStructure.v1_throughAPIGatewayJQuery               |
-        |                     ► http://172.28.0.4/transaction.read.dataList.master.getWorkStructure.v1_throughAPIGatewayJQuery            |
+        | ▪ Call URL        : http(s)://<HOST>/transaction.read.dataList.master.getAll.v1_throughAPIGatewayJQuery               |
+        |                     ► http://172.28.0.4/transaction.read.dataList.master.getAll.v1_throughAPIGatewayJQuery            |
         +--------------------------------------------------------------------------------------------------------------------------+
         | ▪ Version         : 1.0000.0000000                                                                                       |
         | ▪ Last Update     : 2026-08-06                                                                                           |
@@ -74,7 +76,7 @@ namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\rep
         +--------------------------------------------------------------------------------------------------------------------------+
         */
         public function throughAPIGatewayJQuery($varAPIWebToken)
-        {
-        }
+            {
+            }
     }
 }
